Refactor: modularize tray, hotkey, and error handling
