+++ Pointframe.Tests/CountdownWindowTests.cs
