+++ Pointframe.Tests/Models/RecordingSessionGeometryTests.cs
