+++ Pointframe.Tests/Services/Handlers/StaTestHelper.cs
