+++ Pointframe.Tests/Services/OverlayBitmapCaptureTests.cs
