+++ Pointframe.Tests/Services/RecordingCursorEffectsServiceTests.cs
