+++ Pointframe.Tests/Services/ScreenCaptureServiceTests.cs
