+++ Pointframe.Tests/Services/WindowsOcrServiceTests.cs
