+++ Pointframe.Tests/SettingsWindowTests.cs
