+++ Pointframe.Tests/ViewModels/RecordingAnnotationViewModelTests.cs
