+++ Pointframe/App.xaml.cs
@@ -1,66 +1,41 @@
-using System.Diagnostics;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Media.Imaging;
 using System.Windows.Threading;
-using Hardcodet.Wpf.TaskbarNotification;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Pointframe.Automation;
-using Pointframe.Models;
 using Pointframe.Services;
 using Pointframe.Services.Messaging;
 using Pointframe.ViewModels;
 using Serilog;
 using Application = System.Windows.Application;
-using WpfContextMenu = System.Windows.Controls.ContextMenu;
-using WpfMenuItem = System.Windows.Controls.MenuItem;
-using WpfSeparator = System.Windows.Controls.Separator;
 
 namespace Pointframe;
 
 public partial class App : Application
 {
-    private TaskbarIcon? _trayIcon;
     private IHost _host = null!;
     private bool _isAutomationMode;
     private ILogger<App>? _logger;
+    private ILoggerFactory _loggerFactory = null!;
     private IMessageBoxService _messageBox = null!;
     private IUserSettingsService _userSettings = null!;
     private IThemeService _themeService = null!;
     private IAutoUpdateService _autoUpdate = null!;
+    private IDialogService _dialogService = null!;
+    private IImageFileService _imageFileService = null!;
+    private IGlobalHotkeyService _globalHotkey = null!;
+    private IAppErrorHandler _errorHandler = null!;
+    private ITrayIconManager _trayIconManager = null!;
     private IEventSubscription? _updateAvailableSubscription;
     private IEventSubscription? _recordingCompletedSubscription;
     private SettingsWindow? _settingsWindow;
     private AboutWindow? _aboutWindow;
-    private UpdateCheckResult? _pendingUpdate;
-    private string? _pendingRecordingBalloonPath;
-    private WpfMenuItem? _recentRecordingsMenuItem;
-    private readonly List<RecentRecordingItem> _recentRecordings = [];
 
     private const string AutomationOpenImagePathEnvironmentVariable = "SNIPPINGTOOL_AUTOMATION_OPEN_IMAGE_PATH";
-    private const int WH_KEYBOARD_LL = 13;
-    private const int WM_KEYDOWN = 0x0100;
-    private const uint VK_PRINTSCREEN = 0x2C;
-    private const int VK_SHIFT = 0x10;
-
-    private delegate IntPtr LowLevelKeyboardProc(int nCode, IntPtr wParam, IntPtr lParam);
-    private LowLevelKeyboardProc? _keyboardProc; // keep delegate alive to prevent GC
-    private IntPtr _keyboardHook = IntPtr.Zero;
-
-    [DllImport("user32.dll", SetLastError = true)] private static extern IntPtr SetWindowsHookEx(int idHook, LowLevelKeyboardProc lpfn, IntPtr hMod, uint dwThreadId);
-    [DllImport("user32.dll")] private static extern bool UnhookWindowsHookEx(IntPtr hhk);
-    [DllImport("user32.dll")] private static extern IntPtr CallNextHookEx(IntPtr hhk, int nCode, IntPtr wParam, IntPtr lParam);
-    [DllImport("kernel32.dll")] private static extern IntPtr GetModuleHandle(string? lpModuleName);
-    [DllImport("user32.dll")] private static extern short GetAsyncKeyState(int vKey);
-
-    [StructLayout(LayoutKind.Sequential)]
-#pragma warning disable IDE1006 // P/Invoke struct fields must match Windows API names exactly
-    private struct KBDLLHOOKSTRUCT { public uint vkCode, scanCode, flags, time; public IntPtr dwExtraInfo; }
-#pragma warning restore IDE1006
 
     protected override void OnStartup(StartupEventArgs e)
     {
@@ -103,9 +78,14 @@ public partial class App : Application
             .Build();
 
         _logger = _host.Services.GetRequiredService<ILogger<App>>();
+        _loggerFactory = _host.Services.GetRequiredService<ILoggerFactory>();
         _messageBox = _host.Services.GetRequiredService<IMessageBoxService>();
         _userSettings = _host.Services.GetRequiredService<IUserSettingsService>();
         _themeService = _host.Services.GetRequiredService<IThemeService>();
+        _dialogService = _host.Services.GetRequiredService<IDialogService>();
+        _imageFileService = _host.Services.GetRequiredService<IImageFileService>();
+        _globalHotkey = _host.Services.GetRequiredService<IGlobalHotkeyService>();
+        _errorHandler = _host.Services.GetRequiredService<IAppErrorHandler>();
         _themeService.Apply(_userSettings.Current.Theme);
         if (!automationLaunchOptions.IsAutomationMode)
         {
@@ -117,9 +97,7 @@ public partial class App : Application
 
         _logger.LogInformation("Pointframe starting up");
 
-        Current.DispatcherUnhandledException += OnDispatcherUnhandledException;
-        AppDomain.CurrentDomain.UnhandledException += OnAppDomainUnhandledException;
-        TaskScheduler.UnobservedTaskException += OnUnobservedTaskException;
+        _errorHandler.Register();
 
         if (automationLaunchOptions.IsAutomationMode)
         {
@@ -128,12 +106,27 @@ public partial class App : Application
             return;
         }
 
-        InitializeTrayIcon();
-        InitializeRecentRecordingsMenu();
+        _trayIconManager = new TrayIconManager(
+            _host.Services.GetRequiredService<ILogger<TrayIconManager>>(),
+            _messageBox,
+            _host.Services.GetRequiredService<IProcessService>(),
+            _host.Services.GetRequiredService<IUpdateService>(),
+            _host.Services.GetRequiredService<IAppVersionService>(),
+            _autoUpdate,
+            _userSettings,
+            _host.Services.GetRequiredService<IGifExportService>(),
+            onNewSnip: StartSnip,
+            onWholeScreenSnip: StartWholeScreenSnip,
+            onOpenImage: () => Dispatcher.InvokeAsync(OpenImage, System.Windows.Threading.DispatcherPriority.ApplicationIdle),
+            onShowSettings: ShowSettingsWindow,
+            onShowAbout: ShowAboutWindow);
+        _trayIconManager.Initialize();
 #if DEBUG
-        AddDebugMenuItems();
+        _trayIconManager.AddDebugMenuItems();
 #endif
-        RegisterGlobalHotkey();
+        _globalHotkey.RegionSnipRequested += StartSnip;
+        _globalHotkey.WholeScreenSnipRequested += StartWholeScreenSnip;
+        _globalHotkey.Register();
         _logger.LogInformation("Global hotkey registered");
         _host.StartAsync().GetAwaiter().GetResult();
     }
@@ -152,6 +145,8 @@ public partial class App : Application
         services.AddSingleton<IFileSystemService, FileSystemService>();
         services.AddSingleton<IMicrophoneDeviceService, MicrophoneDeviceService>();
         services.AddSingleton<IUserSettingsService, UserSettingsService>();
+        services.AddSingleton<IGlobalHotkeyService, GlobalHotkeyService>();
+        services.AddSingleton<IAppErrorHandler, AppErrorHandler>();
         services.AddTransient<IScreenCaptureService, ScreenCaptureService>();
         services.AddTransient<IVideoWriterFactory, VideoWriterFactory>();
         services.AddTransient<IScreenRecordingService, ScreenRecordingService>();
@@ -160,19 +155,7 @@ public partial class App : Application
         services.AddSingleton<IOcrService, WindowsOcrService>();
         services.AddTransient<OverlayViewModel>();
         services.AddTransient<RecordingAnnotationViewModel>();
-        services.AddTransient<OverlayWindow>(sp => new OverlayWindow(
-            sp.GetRequiredService<OverlayViewModel>(),
-            sp.GetRequiredService<IScreenCaptureService>(),
-            sp.GetRequiredService<IScreenRecordingService>(),
-            sp.GetRequiredService<IMouseHookService>(),
-            sp.GetRequiredService<Func<IScreenRecordingService, string, RecordingHudViewModel>>(),
-            sp.GetRequiredService<IEventAggregator>(),
-            sp.GetRequiredService<ILoggerFactory>(),
-            sp.GetRequiredService<IUserSettingsService>(),
-            sp.GetRequiredService<IMessageBoxService>(),
-            sp.GetRequiredService<IFileSystemService>(),
-            sp.GetRequiredService<IOcrService>(),
-            sp.GetRequiredService<RecordingAnnotationViewModel>()));
+        services.AddTransient<OverlayWindow>(CreateOverlayWindow);
         services.AddTransient<SettingsViewModel>();
         services.AddTransient<SettingsWindow>();
         services.AddTransient<Func<IScreenRecordingService, string, RecordingHudViewModel>>(sp =>
@@ -197,58 +180,31 @@ public partial class App : Application
         services.AddHostedService(sp => sp.GetRequiredService<AutoUpdateService>());
     }
 
+    private static OverlayWindow CreateOverlayWindow(IServiceProvider sp) => new(
+        sp.GetRequiredService<OverlayViewModel>(),
+        sp.GetRequiredService<IScreenCaptureService>(),
+        sp.GetRequiredService<IScreenRecordingService>(),
+        sp.GetRequiredService<IMouseHookService>(),
+        sp.GetRequiredService<Func<IScreenRecordingService, string, RecordingHudViewModel>>(),
+        sp.GetRequiredService<IEventAggregator>(),
+        sp.GetRequiredService<ILoggerFactory>(),
+        sp.GetRequiredService<IUserSettingsService>(),
+        sp.GetRequiredService<IMessageBoxService>(),
+        sp.GetRequiredService<IFileSystemService>(),
+        sp.GetRequiredService<IOcrService>(),
+        sp.GetRequiredService<RecordingAnnotationViewModel>());
+
     protected override void OnExit(ExitEventArgs e)
     {
         _logger?.LogInformation("Pointframe shutting down");
         _updateAvailableSubscription?.Dispose();
         _recordingCompletedSubscription?.Dispose();
-        if (_keyboardHook != IntPtr.Zero)
-        {
-            UnhookWindowsHookEx(_keyboardHook);
-            _keyboardHook = IntPtr.Zero;
-        }
-
-        _trayIcon?.Dispose();
+        _globalHotkey.Dispose();
+        _trayIconManager?.Dispose();
         _host.StopAsync().GetAwaiter().GetResult();
         _host.Dispose();
         base.OnExit(e);
         Log.CloseAndFlush();
-    }
-
-    private void RegisterGlobalHotkey()
-    {
-        _keyboardProc = KeyboardHookCallback;
-        using var process = Process.GetCurrentProcess();
-        var hMod = GetModuleHandle(process.MainModule?.ModuleName);
-        _keyboardHook = SetWindowsHookEx(WH_KEYBOARD_LL, _keyboardProc, hMod, 0);
-        if (_keyboardHook == IntPtr.Zero)
-        {
-            _logger?.LogWarning("Failed to register low-level keyboard hook (error {Code}); Print Screen hotkey will not work",
-                Marshal.GetLastWin32Error());
-        }
-    }
-
-    private IntPtr KeyboardHookCallback(int nCode, IntPtr wParam, IntPtr lParam)
-    {
-        if (nCode >= 0 && wParam == (IntPtr)WM_KEYDOWN)
-        {
-            var kb = Marshal.PtrToStructure<KBDLLHOOKSTRUCT>(lParam);
-            if (kb.vkCode == _userSettings.Current.RegionCaptureHotkey)
-            {
-                if (GetAsyncKeyState(VK_SHIFT) < 0)
-                {
-                    Dispatcher.InvokeAsync(StartWholeScreenSnip);
-                }
-                else
-                {
-                    Dispatcher.InvokeAsync(StartSnip);
-                }
-
-                return (IntPtr)1; // suppress — prevents Windows from handling Print Screen
-            }
-        }
-
-        return CallNextHookEx(_keyboardHook, nCode, wParam, lParam);
     }
 
     private void ShowAutomationWindow(AutomationLaunchOptions automationLaunchOptions)
@@ -286,114 +242,6 @@ public partial class App : Application
         Current.Shutdown();
     }
 
-    private void InitializeTrayIcon()
-    {
-        _trayIcon = new TaskbarIcon
-        {
-            IconSource = new BitmapImage(new Uri("pack://application:,,,/Assets/icon.ico", UriKind.Absolute)),
-            ToolTipText = "Pointframe",
-            ContextMenu = CreateTrayContextMenu(),
-        };
-        _trayIcon.TrayLeftMouseUp += TrayIcon_LeftClick;
-        _trayIcon.TrayBalloonTipClicked += OnTrayBalloonClicked;
-    }
-
-    private WpfContextMenu CreateTrayContextMenu()
-    {
-        var contextMenu = new WpfContextMenu();
-        contextMenu.Items.Add(CreateTrayMenuItem("New Snip", NewSnip_Click));
-        contextMenu.Items.Add(CreateTrayMenuItem("Whole screen snip", WholeScreenSnip_Click));
-        contextMenu.Items.Add(CreateTrayMenuItem("Open image...", OpenImage_Click));
-        contextMenu.Items.Add(new WpfSeparator());
-        contextMenu.Items.Add(CreateTrayMenuItem("Settings", Settings_Click));
-        contextMenu.Items.Add(CreateTrayMenuItem("Check for Updates", CheckForUpdates_Click));
-        contextMenu.Items.Add(CreateTrayMenuItem("About", About_Click));
-        contextMenu.Items.Add(new WpfSeparator());
-        contextMenu.Items.Add(CreateTrayMenuItem("Exit", Exit_Click));
-        return contextMenu;
-    }
-
-    private static WpfMenuItem CreateTrayMenuItem(string header, RoutedEventHandler clickHandler)
-    {
-        var menuItem = new WpfMenuItem
-        {
-            Header = header,
-        };
-        menuItem.Click += clickHandler;
-        return menuItem;
-    }
-
-    private void TrayIcon_LeftClick(object sender, RoutedEventArgs e) => StartSnip();
-    private void NewSnip_Click(object sender, RoutedEventArgs e) => StartSnip();
-    private void WholeScreenSnip_Click(object sender, RoutedEventArgs e) => StartWholeScreenSnip();
-
-    private void InitializeRecentRecordingsMenu()
-    {
-        if (_trayIcon?.ContextMenu is not { } contextMenu)
-        {
-            return;
-        }
-
-        _recentRecordingsMenuItem = new WpfMenuItem
-        {
-            Header = "Recent recordings",
-        };
-
-        contextMenu.Items.Insert(2, _recentRecordingsMenuItem);
-        RebuildRecentRecordingsMenu();
-    }
-
-    private void RebuildRecentRecordingsMenu()
-    {
-        if (_recentRecordingsMenuItem is null)
-        {
-            return;
-        }
-
-        _recentRecordingsMenuItem.Items.Clear();
-
-        if (_recentRecordings.Count == 0)
-        {
-            _recentRecordingsMenuItem.Items.Add(new WpfMenuItem
-            {
-                Header = "No recent recordings",
-                IsEnabled = false,
-            });
-            return;
-        }
-
-        foreach (var recentRecording in _recentRecordings)
-        {
-            var recentRecordingItem = new WpfMenuItem
-            {
-                Header = $"{recentRecording.FileName} ({recentRecording.ElapsedText})",
-            };
-            recentRecordingItem.Items.Add(CreateRecentRecordingActionMenuItem("Open", OpenRecentRecording_Click, recentRecording));
-            recentRecordingItem.Items.Add(CreateRecentRecordingActionMenuItem("Open folder", OpenRecentRecordingFolder_Click, recentRecording));
-            recentRecordingItem.Items.Add(CreateRecentRecordingActionMenuItem("Export to GIF", ExportRecentRecordingGif_Click, recentRecording));
-            _recentRecordingsMenuItem.Items.Add(recentRecordingItem);
-        }
-    }
-
-    private static WpfMenuItem CreateRecentRecordingActionMenuItem(
-        string header,
-        RoutedEventHandler clickHandler,
-        RecentRecordingItem recentRecording)
-    {
-        var menuItem = new WpfMenuItem
-        {
-            Header = header,
-            Tag = recentRecording,
-        };
-        menuItem.Click += clickHandler;
-        return menuItem;
-    }
-
-    private void OpenImage_Click(object sender, RoutedEventArgs e)
-    {
-        Dispatcher.InvokeAsync(OpenImage, DispatcherPriority.ApplicationIdle);
-    }
-
     private void OpenImage()
     {
         var selectedPath = _isAutomationMode
@@ -401,8 +249,7 @@ public partial class App : Application
             : null;
         if (string.IsNullOrWhiteSpace(selectedPath))
         {
-            var dialogService = _host.Services.GetRequiredService<IDialogService>();
-            selectedPath = dialogService.PickOpenImageFile();
+            selectedPath = _dialogService.PickOpenImageFile();
         }
 
         if (string.IsNullOrWhiteSpace(selectedPath))
@@ -412,8 +259,7 @@ public partial class App : Application
 
         try
         {
-            var imageFileService = _host.Services.GetRequiredService<IImageFileService>();
-            var bitmap = imageFileService.LoadForAnnotation(selectedPath);
+            var bitmap = _imageFileService.LoadForAnnotation(selectedPath);
             ShowOverlayFromImage(bitmap, selectedPath);
         }
         catch (Exception ex) when (ex is FileNotFoundException or InvalidDataException or NotSupportedException or IOException or UnauthorizedAccessException)
@@ -430,69 +276,6 @@ public partial class App : Application
         }
     }
 
-    private void Exit_Click(object sender, RoutedEventArgs e) => Current.Shutdown();
-
-#if DEBUG
-    private void AddDebugMenuItems()
-    {
-        if (_trayIcon?.ContextMenu is not { } contextMenu)
-        {
-            return;
-        }
-
-        var simulateUiErrorMenuItem = new System.Windows.Controls.MenuItem
-        {
-            Header = "Simulate UI Error",
-            InputGestureText = "Ctrl+Shift+F12"
-        };
-        simulateUiErrorMenuItem.Click += SimulateUiError_Click;
-
-        contextMenu.Items.Insert(Math.Max(0, contextMenu.Items.Count - 1), simulateUiErrorMenuItem);
-    }
-
-    private void SimulateUiError_Click(object sender, RoutedEventArgs e)
-    {
-        _logger?.LogDebug("Simulating UI recovery smoke test from tray menu");
-        throw new InvalidOperationException("Debug-only UI recovery smoke test.");
-    }
-#endif
-
-    private async void CheckForUpdates_Click(object sender, RoutedEventArgs e)
-    {
-        var menuItem = (System.Windows.Controls.MenuItem)sender;
-        menuItem.IsEnabled = false;
-
-        try
-        {
-            var updateService = _host.Services.GetRequiredService<IUpdateService>();
-            var result = await updateService.CheckForUpdates();
-
-            if (!result.IsUpdateAvailable)
-            {
-                var current = _host.Services.GetRequiredService<IAppVersionService>().Current;
-                _messageBox.ShowInformation(
-                    $"You're already on the latest version (v{current.Major}.{current.Minor}.{current.Build}).",
-                    "Check for Updates");
-                return;
-            }
-
-            await _autoUpdate.ConfirmAndInstall(result);
-        }
-        catch (Exception ex)
-        {
-            _logger?.LogError(ex, "Update check failed");
-            _messageBox.ShowWarning(
-                "Could not check for updates. Please check your internet connection and try again.",
-                "Check for Updates");
-        }
-        finally
-        {
-            menuItem.IsEnabled = true;
-        }
-    }
-
-    private void Settings_Click(object sender, RoutedEventArgs e) => ShowSettingsWindow();
-
     private void ShowSettingsWindow()
     {
         if (_settingsWindow is not null)
@@ -506,8 +289,6 @@ public partial class App : Application
         _settingsWindow.Closed += (_, _) => _settingsWindow = null;
         _settingsWindow.Show();
     }
-
-    private void About_Click(object sender, RoutedEventArgs e) => ShowAboutWindow();
 
     private void ShowAboutWindow()
     {
@@ -540,7 +321,7 @@ public partial class App : Application
     private void ShowAutomationTraySampleOverlayWindow()
     {
         AutomationSampleFactory.CreateOpenedImageSample();
-        OpenImage_Click(this, new RoutedEventArgs());
+        OpenImage();
     }
 
     private void ShowOverlayFromImage(BitmapSource bitmap, string sourcePath)
@@ -596,7 +377,7 @@ public partial class App : Application
     private void StartCapture(SnipLaunchMode launchMode)
     {
         _logger?.LogDebug("{Mode} snip started", launchMode);
-        var delay = _host.Services.GetRequiredService<IUserSettingsService>().Current.CaptureDelaySeconds;
+        var delay = _userSettings.Current.CaptureDelaySeconds;
         if (delay > 0)
         {
             new CountdownWindow(delay, () => ShowSelectionOverlay(launchMode)).Show();
@@ -609,10 +390,9 @@ public partial class App : Application
     private async void ShowSelectionOverlay(SnipLaunchMode launchMode)
     {
         var screenCapture = _host.Services.GetRequiredService<IScreenCaptureService>();
-        var loggerFactory = _host.Services.GetRequiredService<ILoggerFactory>();
         var selection = launchMode == SnipLaunchMode.WholeScreen
-            ? await SelectionSession.SelectWholeScreenAsync(screenCapture, loggerFactory)
-            : await SelectionSession.SelectAsync(screenCapture, loggerFactory);
+            ? await SelectionSession.SelectWholeScreenAsync(screenCapture, _loggerFactory)
+            : await SelectionSession.SelectAsync(screenCapture, _loggerFactory);
 
         if (selection is null)
         {
@@ -624,237 +404,16 @@ public partial class App : Application
         DpiAwarenessScope.RunPerMonitorV2(() => overlay.Show());
     }
 
-    private void ShowUpdateBalloon(UpdateCheckResult result)
-    {
-        _pendingRecordingBalloonPath = null;
-        var v = result.LatestVersion;
-        _trayIcon?.ShowBalloonTip(
-            "Update Available",
-            $"Version {v.Major}.{v.Minor}.{v.Build} is ready to download.",
-            BalloonIcon.Info);
-    }
-
     private ValueTask HandleUpdateAvailable(UpdateAvailableMessage message)
     {
-        _pendingUpdate = message.Result;
-        ShowUpdateBalloon(message.Result);
+        _trayIconManager.HandleUpdateAvailable(message.Result);
         return ValueTask.CompletedTask;
     }
 
     private ValueTask HandleRecordingCompleted(RecordingCompletedMessage message)
     {
-        var recentRecording = new RecentRecordingItem(message.OutputPath, message.ElapsedText);
-        _recentRecordings.RemoveAll(item => string.Equals(item.OutputPath, recentRecording.OutputPath, StringComparison.OrdinalIgnoreCase));
-        _recentRecordings.Insert(0, recentRecording);
-        if (_recentRecordings.Count > 5)
-        {
-            _recentRecordings.RemoveRange(5, _recentRecordings.Count - 5);
-        }
-
-        RebuildRecentRecordingsMenu();
-        ShowRecordingCompletedBalloon(recentRecording);
+        _trayIconManager.HandleRecordingCompleted(message.OutputPath, message.ElapsedText);
         return ValueTask.CompletedTask;
-    }
-
-    private void ShowRecordingCompletedBalloon(RecentRecordingItem recentRecording)
-    {
-        _pendingRecordingBalloonPath = recentRecording.OutputPath;
-        var directory = Path.GetDirectoryName(recentRecording.OutputPath) ?? recentRecording.OutputPath;
-        _trayIcon?.ShowBalloonTip(
-            "Recording saved",
-            $"{recentRecording.FileName} • {recentRecording.ElapsedText}{Environment.NewLine}{directory}",
-            BalloonIcon.Info);
-    }
-
-    private async void OnTrayBalloonClicked(object sender, System.Windows.RoutedEventArgs e)
-    {
-        if (_pendingUpdate is not null)
-        {
-            var update = _pendingUpdate;
-            _pendingUpdate = null;
-            await _autoUpdate.ConfirmAndInstall(update);
-            return;
-        }
-
-        if (string.IsNullOrWhiteSpace(_pendingRecordingBalloonPath))
-        {
-            return;
-        }
-
-        OpenPath(_pendingRecordingBalloonPath);
-        _pendingRecordingBalloonPath = null;
-    }
-
-    private void OpenRecentRecording_Click(object sender, RoutedEventArgs e)
-    {
-        if (sender is not WpfMenuItem { Tag: RecentRecordingItem recentRecording })
-        {
-            return;
-        }
-
-        OpenPath(recentRecording.OutputPath);
-    }
-
-    private void OpenRecentRecordingFolder_Click(object sender, RoutedEventArgs e)
-    {
-        if (sender is not WpfMenuItem { Tag: RecentRecordingItem recentRecording })
-        {
-            return;
-        }
-
-        var directory = Path.GetDirectoryName(recentRecording.OutputPath);
-        if (!string.IsNullOrWhiteSpace(directory))
-        {
-            OpenFolder(directory);
-        }
-    }
-
-    private async void ExportRecentRecordingGif_Click(object sender, RoutedEventArgs e)
-    {
-        if (sender is not WpfMenuItem { Tag: RecentRecordingItem recentRecording } menuItem)
-        {
-            return;
-        }
-
-        if (!File.Exists(recentRecording.OutputPath))
-        {
-            _messageBox.ShowWarning("The recording file could not be found.", "Export to GIF");
-            return;
-        }
-
-        var gifPath = Path.ChangeExtension(recentRecording.OutputPath, ".gif");
-        menuItem.IsEnabled = false;
-
-        try
-        {
-            var gifExportService = _host.Services.GetRequiredService<IGifExportService>();
-            await gifExportService.Export(recentRecording.OutputPath, gifPath, _userSettings.Current.GifFps).ConfigureAwait(true);
-            var directory = Path.GetDirectoryName(gifPath) ?? gifPath;
-            _trayIcon?.ShowBalloonTip(
-                "GIF exported",
-                $"{Path.GetFileName(gifPath)} is ready.{Environment.NewLine}{directory}",
-                BalloonIcon.Info);
-        }
-        catch (Exception ex)
-        {
-            _logger?.LogError(ex, "GIF export from recent recordings failed for {Path}", recentRecording.OutputPath);
-            _messageBox.ShowWarning("The GIF export failed. Please try again.", "Export to GIF");
-        }
-        finally
-        {
-            menuItem.IsEnabled = true;
-        }
-    }
-
-    private void OpenPath(string path)
-    {
-        if (!File.Exists(path))
-        {
-            _messageBox.ShowWarning("The selected recording file could not be found.", "Open Recording");
-            return;
-        }
-
-        var processService = _host.Services.GetRequiredService<IProcessService>();
-        processService.Start(new ProcessStartInfo(path)
-        {
-            UseShellExecute = true,
-        });
-    }
-
-    private void OpenFolder(string path)
-    {
-        var processService = _host.Services.GetRequiredService<IProcessService>();
-        processService.Start(new ProcessStartInfo("explorer.exe", path));
-    }
-
-    private sealed record RecentRecordingItem(string OutputPath, string ElapsedText)
-    {
-        public string FileName => Path.GetFileName(OutputPath);
-    }
-
-    private void OnDispatcherUnhandledException(object sender, System.Windows.Threading.DispatcherUnhandledExceptionEventArgs e)
-    {
-        _logger?.LogError(e.Exception, "Unhandled dispatcher exception");
-        e.Handled = true;
-
-        var closedWindowName = TryRecoverFromActiveWindow();
-        var recoveryMessage = closedWindowName is null
-            ? "You can continue using Pointframe. Details have been written to the log file."
-            : $"{closedWindowName} was closed so Pointframe can recover. You can continue using the app. Details have been written to the log file.";
-
-        _messageBox.ShowError(
-            $"Something went wrong while processing your last action.\n\n{e.Exception.Message}\n\n{recoveryMessage}",
-            "Pointframe — Recovered From Error");
-    }
-
-    private string? TryRecoverFromActiveWindow()
-    {
-        try
-        {
-            var window = GetRecoveryWindow();
-            if (window is null)
-            {
-                return null;
-            }
-
-            var windowName = string.IsNullOrWhiteSpace(window.Title)
-                ? window.GetType().Name
-                : window.Title;
-
-            CloseWindowTree(window);
-            _logger?.LogWarning(
-                "Closed window {WindowType} during dispatcher exception recovery",
-                window.GetType().Name);
-
-            return windowName;
-        }
-        catch (Exception recoveryException)
-        {
-            _logger?.LogError(recoveryException, "Failed to recover active window after dispatcher exception");
-            return null;
-        }
-    }
-
-    private Window? GetRecoveryWindow()
-    {
-        var visibleWindows = Current.Windows
-            .OfType<Window>()
-            .Where(window => window.IsVisible)
-            .ToList();
-
-        return visibleWindows.FirstOrDefault(window => window.IsActive)
-            ?? visibleWindows.FirstOrDefault(window => window is OverlayWindow
-                or CountdownWindow
-                or UpdateDownloadWindow
-                or SettingsWindow
-                or AboutWindow
-                or PinnedScreenshotWindow)
-            ?? visibleWindows.FirstOrDefault();
-    }
-
-    private static void CloseWindowTree(Window rootWindow)
-    {
-        foreach (var ownedWindow in rootWindow.OwnedWindows.OfType<Window>().ToList())
-        {
-            CloseWindowTree(ownedWindow);
-        }
-
-        if (rootWindow.IsVisible)
-        {
-            rootWindow.Close();
-        }
-    }
-
-    private void OnAppDomainUnhandledException(object sender, UnhandledExceptionEventArgs e)
-    {
-        _logger?.LogCritical(e.ExceptionObject as Exception, "Unhandled AppDomain exception — application will terminate");
-        Log.CloseAndFlush();
-    }
-
-    private void OnUnobservedTaskException(object? sender, UnobservedTaskExceptionEventArgs e)
-    {
-        _logger?.LogError(e.Exception, "Unobserved task exception");
-        e.SetObserved();
     }
 }
