+++ Pointframe/Services/GlobalHotkeyService.cs
@@ -1,0 +1,77 @@
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+using Microsoft.Extensions.Logging;
+using WpfApplication = System.Windows.Application;
+
+namespace Pointframe.Services;
+
+internal sealed class GlobalHotkeyService : IGlobalHotkeyService
+{
+    private readonly IUserSettingsService _userSettings;
+    private readonly ILogger<GlobalHotkeyService> _logger;
+
+    private NativeMethods.LowLevelKeyboardProc? _keyboardProc;
+    private IntPtr _keyboardHook = IntPtr.Zero;
+    private bool _disposed;
+
+    public event Action? RegionSnipRequested;
+    public event Action? WholeScreenSnipRequested;
+
+    public GlobalHotkeyService(IUserSettingsService userSettings, ILogger<GlobalHotkeyService> logger)
+    {
+        _userSettings = userSettings;
+        _logger = logger;
+    }
+
+    public void Register()
+    {
+        _keyboardProc = HookCallback;
+        using var process = Process.GetCurrentProcess();
+        var hMod = NativeMethods.GetModuleHandle(process.MainModule?.ModuleName);
+        _keyboardHook = NativeMethods.SetWindowsHookEx(NativeMethods.WH_KEYBOARD_LL, _keyboardProc, hMod, 0);
+        if (_keyboardHook == IntPtr.Zero)
+        {
+            _logger.LogWarning(
+                "Failed to register low-level keyboard hook (error {Code}); Print Screen hotkey will not work",
+                Marshal.GetLastWin32Error());
+        }
+    }
+
+    private IntPtr HookCallback(int nCode, IntPtr wParam, IntPtr lParam)
+    {
+        if (nCode >= 0 && wParam == (IntPtr)NativeMethods.WM_KEYDOWN)
+        {
+            var kb = Marshal.PtrToStructure<NativeMethods.KBDLLHOOKSTRUCT>(lParam);
+            if (kb.vkCode == _userSettings.Current.RegionCaptureHotkey)
+            {
+                if (NativeMethods.GetAsyncKeyState(NativeMethods.VK_SHIFT) < 0)
+                {
+                    WpfApplication.Current.Dispatcher.InvokeAsync(() => WholeScreenSnipRequested?.Invoke());
+                }
+                else
+                {
+                    WpfApplication.Current.Dispatcher.InvokeAsync(() => RegionSnipRequested?.Invoke());
+                }
+
+                return (IntPtr)1;
+            }
+        }
+
+        return NativeMethods.CallNextHookEx(_keyboardHook, nCode, wParam, lParam);
+    }
+
+    public void Dispose()
+    {
+        if (_disposed)
+        {
+            return;
+        }
+
+        _disposed = true;
+        if (_keyboardHook != IntPtr.Zero)
+        {
+            NativeMethods.UnhookWindowsHookEx(_keyboardHook);
+            _keyboardHook = IntPtr.Zero;
+        }
+    }
+}
