+++ Pointframe/Services/IAppErrorHandler.cs
@@ -1,0 +1,6 @@
+namespace Pointframe.Services;
+
+internal interface IAppErrorHandler
+{
+    void Register();
+}
