+++ Pointframe/Services/IGlobalHotkeyService.cs
@@ -1,0 +1,10 @@
+namespace Pointframe.Services;
+
+internal interface IGlobalHotkeyService : IDisposable
+{
+    event Action RegionSnipRequested;
+
+    event Action WholeScreenSnipRequested;
+
+    void Register();
+}
