+++ Pointframe/Services/ITrayIconManager.cs
@@ -1,0 +1,12 @@
+namespace Pointframe.Services;
+
+internal interface ITrayIconManager : IDisposable
+{
+    void Initialize();
+
+    void AddDebugMenuItems();
+
+    void HandleUpdateAvailable(Models.UpdateCheckResult result);
+
+    void HandleRecordingCompleted(string outputPath, string elapsedText);
+}
